requirements.txt
fastapi>=0.115,<1
gliner>=0.2.24,<0.3
httpx>=0.28,<1
uvicorn[standard]>=0.34,<1
google-genai>=1.73.1,<2
langchain>=1.2.15,<2
langgraph>=1.1.6,<2
langgraph-checkpoint-sqlite>=3.0.3,<4
langchain-google-genai>=4.2.2,<5
langchain-ollama>=0.3,<1
langchain-mcp-adapters>=0.2.2,<1
mcp[cli]>=1.12,<2
psycopg[binary]>=3.2,<4
qdrant-client>=1.12,<2
openpyxl>=3.1,<4
pypdf>=5.0,<6
